Reorganize some files and classes
