+++ CCVTAC/CCVTAC.Console/Downloading/DownloadEntities/DownloadEntityFactory.cs
@@ -1,6 +1,6 @@
 using System.Text.RegularExpressions;
 
-namespace CCVTAC.Console.DownloadEntities;
+namespace CCVTAC.Console.Downloading.DownloadEntities;
 
 public static class DownloadEntityFactory
 {
