+++ CCVTAC/CCVTAC.Console/Downloading/DownloadEntities/IDownloadEntity.cs
@@ -1,6 +1,6 @@
 using System.Text.RegularExpressions;
 
-namespace CCVTAC.Console.DownloadEntities;
+namespace CCVTAC.Console.Downloading.DownloadEntities;
 
 public interface IDownloadEntity
 {
